helplineApp/requirements.txt
coverage==4.5.3
Django==2.2.1
django-filter==2.1.0
django-nose==1.4.6
djangorestframework==3.9.4
djangorestframework-jwt==1.11.0
Markdown==3.1.1
nose==1.3.7
PyJWT==1.7.1
pytz==2019.1
sqlparse==0.3.0
